# frontend/src/shared/api/generated/campaign-manager.openapi.json
{
  "components": {
    "schemas": {
...
      "AudienceEstimateResponse": {
        "properties": {
          "estimated_users": {
            "title": "Estimated Users",
            "type": "integer"
          },
          "rfm_filters": {
            "additionalProperties": true,
            "title": "Rfm Filters",
            "type": "object"
          },
          "segments": {
            "items": {
              "type": "integer"
            },
            "title": "Segments",
            "type": "array"
          }
        },
        "required": [
          "estimated_users",
          "segments",
          "rfm_filters"
        ],
        "title": "AudienceEstimateResponse",
        "type": "object"
      },
...
      "CampaignSummary": {
        "properties": {
          "budget_spent": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Budget Spent",
            "type": "string"
          },
          "budget_total": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Budget Total",
            "type": "string"
          },
          "campaign_id": {
            "format": "uuid",
            "title": "Campaign Id",
            "type": "string"
          },
          "cashback_rate": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Cashback Rate",
            "type": "string"
          },
          "end_date": {
            "format": "date-time",
            "title": "End Date",
            "type": "string"
          },
          "mcc_codes": {
            "items": {
              "type": "string"
            },
            "title": "Mcc Codes",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "start_date": {
            "format": "date-time",
            "title": "Start Date",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "campaign_id",
          "name",
          "status",
          "budget_total",
          "budget_spent",
          "cashback_rate",
          "start_date",
          "end_date"
        ],
        "title": "CampaignSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Campaign CRUD, applicability filtering, A/B testing and analytics for the CashBack stack.",
    "title": "Cashback Campaign Manager",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/campaigns": {
      "get": {
        "description": "Full campaign list for the admin UI (any status, newest first).",
        "operationId": "list_campaigns_campaigns_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(DRAFT|ACTIVE|PAUSED|COMPLETED)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "request",
            "required": true,
            "schema": {
              "title": "Request"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CampaignResponse"
                  },
                  "title": "Response List Campaigns Campaigns Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Campaigns",
        "tags": [
          "campaigns"
        ]
      },
      "post": {
        "operationId": "create_campaign_campaigns_post",
        "parameters": [
          {
            "in": "query",
            "name": "request",
            "required": true,
            "schema": {
              "title": "Request"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Campaign",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/campaigns/active": {
      "get": {
        "operationId": "list_active_campaigns_campaigns_active_get",
        "parameters": [
          {
            "in": "query",
            "name": "request",
            "required": true,
            "schema": {
              "title": "Request"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CampaignSummary"
                  },
                  "title": "Response List Active Campaigns Campaigns Active Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Active Campaigns",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/campaigns/applicable/{user_id}": {
      "get": {
        "operationId": "applicable_for_user_campaigns_applicable__user_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "request",
            "required": true,
            "schema": {
              "title": "Request"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Applicable For User Campaigns Applicable  User Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Applicable For User",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/campaigns/audience/estimate": {
      "get": {
        "operationId": "audience_estimate_campaigns_audience_estimate_get",
        "parameters": [
          {
            "in": "query",
            "name": "segment_ids",
            "required": true,
            "schema": {
              "items": {
                "type": "integer"
              },
              "minItems": 1,
              "title": "Segment Ids",
              "type": "array"
            }
          },
          {
            "description": "recency_days <= rfmR",
            "in": "query",
            "name": "rfmR",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "recency_days <= rfmR",
              "title": "Rfmr"
            }
          },
          {
            "description": "frequency_total >= rfmF",
            "in": "query",
            "name": "rfmF",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "frequency_total >= rfmF",
              "title": "Rfmf"
            }
          },
          {
            "description": "monetary_total >= rfmM",
            "in": "query",
            "name": "rfmM",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "monetary_total >= rfmM",
              "title": "Rfmm"
            }
          },
          {
            "in": "query",
            "name": "request",
            "required": true,
            "schema": {
              "title": "Request"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudienceEstimateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Audience Estimate",
        "tags": [
          "campaigns"
        ]
      }
    },
...
    "/campaigns/{campaign_id}/budget/check": {
      "post": {
        "operationId": "reserve_budget_campaigns__campaign_id__budget_check_post",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Campaign Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "request",
            "required": true,
            "schema": {
              "title": "Request"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BudgetCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetCheckResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reserve Budget",
        "tags": [
          "campaigns"
        ]
      }
    },
...
  }
}